requirements.txt
numpy
torch
torchvision
gsplat
omegaconf
PyYAML
tqdm
Pillow
ninja
jaxtyping
rich
packaging
setuptools
wheel
